requirements.txt
markdown==3.6
pypandoc==1.15
pillow==10.2.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
cssutils==2.11.1
dateparser==1.2.1
feedparser==6.0.11
html2text==2025.4.15
htmldate==1.9.3
idna==3.10
lxml==5.4.0
pandas==2.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
regex==2024.11.6
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.4.0
supabase==1.0.3
python-dotenv==1.0.0
jinja2==3.1.2
